refactor(branch): split policy, transaction, delta execution, and reporting into focused modules
The PR #1046 review flagged ablate.py (1,203→1,320 lines) and
rollout_branch.py (1,294→1,320) as monoliths, branch() at cyclomatic 21+
with 75 statements, and _run_children() taking 14 arguments. Behavior-
neutral four-way split along the reviewed seams:

- branch_policy.py — stage/snapshot policy: layer resolution + capability
  gating, capture_stage, recorded-stage resolution, and the whole
  delta-vector gate (validate_deltas, the per-delta stage/layer/parent/
  runner preconditions, the fresh-children boundary rule).
- branch_transaction.py — the transactional heart: checkpoint_parent, the
  scoped LinearState capture, and BranchTransaction — one dataclass
  carrying what _run_children took as 14 arguments, with the per-child
  restore/run/record/hand-off loop as methods.
- branch_children.py — delta execution paths: the in-place default runner
  and per-child runner selection; the fresh-rollout half stays implemented
  in branch_skill.py (its import path and run_fresh_child patch seam are
  pinned by six test files) and is re-exported here.
- branch_report.py — the ablation report model, arm/child pairing,
  per-test mining and attribution, moved out of ablate.py.

rollout_branch.py keeps the branch() orchestrator — now cc≈15 / 41
statements (was cc≈30 / 72) — plus the failure-isolated lineage write,
which stays because tests patch benchflow.rollout_branch.write_branch_artifacts;
the transaction's fresh-runner factory and in-place result writer are
injected from rollout_branch's globals for the same reason
(tests/test_branch_child_result.py patches them there). ablate.py lands
at 884 lines and rollout_branch.py at 441; every previously importable
name keeps working via re-exports (rollout_branch re-exports the gates,
exceptions, capture_stage, ChildRunner, CHILD_WALL_CLOCK_KEY,
_LinearState and the fresh-child API; ablate re-exports the report and
attribution API). Full suite identical before and after (6129 passed);
no test edited.
