+++ src/benchflow/rollout_branch.py
@@ -1,33 +1,36 @@
-"""The Branch -> Rollout engine wiring.
+"""The Branch -> Rollout engine wiring — the fork orchestrator.
 
 The pure Branch primitives live in :mod:`benchflow.branch` — ``checkpoint``,
 ``restore``, ``aggregate`` operate on a ``RolloutTree`` node and an
 ``Environment`` with no I/O beyond the env contract. This module is the
-*engine*: it drives those primitives against a live
-:class:`~benchflow.rollout.Rollout` — quiescing the agent, running each forked
-child as an **isolated sub-rollout**, and restoring the parent's linear state
-afterward.
+*orchestrator*: :func:`branch` drives one fork end to end against a live
+:class:`~benchflow.rollout.Rollout` — admit the request, quiesce the agent,
+checkpoint, run the children, restore the parent, aggregate, leave lineage —
+delegating each concern to its focused module:
 
-Why a separate module: ``rollout.py`` is the 5-phase lifecycle; the Branch
-path is a distinct, optional capability. Keeping it here holds ``rollout.py``
-under the size threshold and keeps the branch logic independently testable.
+* :mod:`benchflow.branch_policy` — what may be captured and forked, and when:
+  layer resolution and capability gating, the stage-snapshot capture path
+  (:func:`~benchflow.branch_policy.capture_stage`), recorded-stage resolution,
+  and the whole delta-vector gate. Everything fails closed before anything is
+  quiesced.
+* :mod:`benchflow.branch_transaction` — the transactional heart: the composed
+  checkpoint at the branch point, the scoped
+  :class:`~benchflow.branch_transaction.LinearState` capture that makes each
+  child an isolated sub-rollout, and the per-child restore/run/record loop
+  (:class:`~benchflow.branch_transaction.BranchTransaction`).
+* :mod:`benchflow.branch_children` — how a child executes its delta: the
+  in-place default runner, and (via :mod:`benchflow.branch_skill`) the
+  fresh-rollout path every ``env-ready`` child takes.
+* :mod:`benchflow.branch_report` — lineage and ablation reporting; the
+  fork's failure-isolated lineage write stays here (:func:`_write_lineage`)
+  because its writers must resolve through *this* module's namespace — the
+  patch seam ``tests/test_rollout_branch.py`` pins.
 
-The engine functions are free functions taking a ``Rollout`` as their first
-argument — ``Rollout.branch`` / ``Rollout.mark_stage`` /
-``Rollout.branch_at_stage`` are thin entry points that delegate here.
-:func:`capture_stage` is the stage-boundary policy's one capture path (RFC
-§3.2): the lifecycle calls it at the boundaries named in
-``RolloutConfig.snapshot_stages``, and :func:`branch` with ``at_stage=`` forks
-from what it recorded.
-
-Most children run *in place* on the parent Rollout (a fresh agent session over
-the restored world). Children forked from ``env-ready`` cannot: that boundary
-precedes ``install_agent()``, so the restored world has no agent to connect to
-and none of the skills, sandbox user or lockdown installation deploys — each
-child re-runs installation as its own Rollout over the restored sandbox,
-whatever its delta. This module keeps the gates
-(:func:`_validate_skill_delta`, :func:`_validate_fresh_children`);
-:mod:`benchflow.branch_skill` owns that child.
+``Rollout.branch`` / ``Rollout.mark_stage`` / ``Rollout.branch_at_stage`` are
+thin entry points that delegate here, and the public names the engine has
+always exported (``ChildRunner``, ``CHILD_WALL_CLOCK_KEY``, the gate
+exceptions, ``capture_stage``, ``make_default_runner``, the fresh-child API)
+are re-exported so existing import paths keep working.
 
 Isolation invariant (the architecture's "tree is additive / no-regression"):
 after :func:`branch` returns, the parent Rollout's linear state — ``_cursor``,
@@ -35,687 +38,77 @@ after :func:`branch` returns, the parent Rollout's linear state — ``_cursor``,
 bookkeeping) — is *exactly* what it was before, and so is the result-bearing
 state an in-place child writes on its way through (``_timing``,
 ``_verifier_error``, ``_diagnostics``, the usage counters — see
-:data:`_RESULT_STATE_FIELDS`): the parent's own ``result.json`` describes the
-parent's run, never the last arm's. A branch child never re-entrantly mutates
-the shared instance: it runs against a scoped snapshot of that state, captured
-before and restored after each child, and its real continuation Steps attach to
-a *pending* branch-child node so the reward and value land on the right node.
+:data:`~benchflow.branch_transaction._RESULT_STATE_FIELDS`): the parent's own
+``result.json`` describes the parent's run, never the last arm's. A branch
+child never re-entrantly mutates the shared instance: it runs against a scoped
+snapshot of that state, captured before and restored after each child, and its
+real continuation Steps attach to a *pending* branch-child node so the reward
+and value land on the right node.
 """
 
 from __future__ import annotations
 
-import copy
-import dataclasses
 import logging
-import time
-from collections.abc import Awaitable, Callable, Iterable, Sequence
-from dataclasses import dataclass
-from datetime import datetime
+from collections.abc import Sequence
 from typing import TYPE_CHECKING, Any
 
-from benchflow._utils.config_override import validate_overlay
-from benchflow.branch import UNSCORED_KEY, StageSnapshot, UnscoredChildError
-from benchflow.branch import adopt_checkpoint as _adopt_checkpoint
-from benchflow.branch import aggregate as _aggregate_branch
-from benchflow.branch import checkpoint as _checkpoint_branch
-from benchflow.branch import checkpoint_composed as _checkpoint_composed
-from benchflow.branch import restore as _restore_branch
-from benchflow.branch import restore_composed as _restore_composed
-from benchflow.branch_artifacts import MountedArtifacts, child_mount_dir
-from benchflow.branch_delta import BranchDelta, BranchDeltaNotSupported
-from benchflow.branch_lineage import write_branch_artifacts, write_stage_snapshots
-from benchflow.branch_result import (
-    scope_child_result_state,
-    write_in_place_child_result,
+from benchflow.branch import (  # noqa: F401  (StageSnapshot/UnscoredChildError re-exported)
+    UNSCORED_KEY,
+    StageSnapshot,
+    UnscoredChildError,
 )
-from benchflow.branch_skill import (
+from benchflow.branch import aggregate as _aggregate_branch
+from benchflow.branch_artifacts import MountedArtifacts, child_mount_dir  # noqa: F401
+from benchflow.branch_children import (  # noqa: F401  (re-exports)
     EXECUTION_FRESH_ROLLOUT,
     FRESH_CHILD_LAYER,
     FRESH_CHILD_STAGE,
     SKILL_DELTA_LAYER,
     SKILL_DELTA_STAGE,
+    ChildRunner,
+    make_default_runner,
     make_fresh_child_runner,
     resolve_child_skill_policy,
     resolve_environment_ref_delta,
 )
-from benchflow.branch_stage import (
-    BranchStageNotCaptured,
-    captured_stages,
-    validate_stage,
+from benchflow.branch_delta import BranchDelta, BranchDeltaNotSupported  # noqa: F401
+from benchflow.branch_lineage import (
+    write_branch_artifacts,
+    write_stage_snapshots,
 )
-from benchflow.models import TrajectorySource
-from benchflow.skill_policy import SKILL_MODE_NO_SKILL
-from benchflow.trajectories.tree import RolloutNode
+from benchflow.branch_policy import (  # noqa: F401  (re-exports)
+    _EXECUTABLE_DELTA_FIELDS,
+    _UNSUPPORTED_DELTA_FIELDS,
+    BranchChildExecutionNotSupported,
+    BranchParentSkillModeConflict,
+    capture_stage,
+    gate_layers,
+    recorded_stage_checkpoint,
+    resolve_layers,
+    runs_fresh_children,
+    validate_deltas,
+    validate_fresh_children,
+)
+from benchflow.branch_result import (  # noqa: F401  (scope_… re-exported)
+    scope_child_result_state,
+    write_in_place_child_result,
+)
+from benchflow.branch_transaction import (  # noqa: F401  (wall-clock key re-exported)
+    CHILD_WALL_CLOCK_KEY,
+    BranchTransaction,
+    LinearState,
+    checkpoint_parent,
+)
 
 if TYPE_CHECKING:
     from benchflow.rollout import Rollout
+    from benchflow.trajectories.tree import RolloutNode
 
 logger = logging.getLogger(__name__)
 
-# The checkpoint layers a branch can compose (RFC §3.1). Agent-session is
-# layer three and explicitly out of scope for v1.
-_SNAPSHOT_LAYERS = frozenset({"environment", "sandbox"})
-
-#: Node-state key holding how long a branch child took, in seconds. Recorded
-#: on the child node (even when the child raised) so a caller comparing
-#: children — ``bench eval ablate``'s per-arm cost column — can attribute time
-#: without re-deriving it from per-child artifacts, which only a
-#: fresh-rollout child leaves behind. Deliberately *not* serialized by
-#: :mod:`benchflow.branch_lineage`: a measured duration would make
-#: ``tree.json`` non-deterministic.
-CHILD_WALL_CLOCK_KEY = "wall_clock_sec"
-
-# The BranchDelta fields the engine executes vs records-only.
-# ``injected_prompt`` runs in place on the parent's rollout; ``skill_mode``,
-# ``config_override`` and ``environment_ref`` run the child as a fresh rollout
-# from the env-ready snapshot (RFC §3.3, gated by
-# :func:`_validate_skill_delta` / :func:`_validate_config_delta` /
-# :func:`_validate_environment_delta`). Any remaining field exists in the
-# schema and provenance so the artifact format is stable, and fails closed
-# here. Derived from the schema (all BranchDelta fields minus the executable
-# set) so a future BranchDelta field is unsupported-by-default — it fails
-# closed here instead of being silently ignored. Sorted for deterministic
-# errors; empty today, and load-bearing the day a field is added.
-_EXECUTABLE_DELTA_FIELDS = frozenset(
-    {"injected_prompt", "skill_mode", "config_override", "environment_ref"}
-)
-_UNSUPPORTED_DELTA_FIELDS = tuple(
-    sorted({f.name for f in dataclasses.fields(BranchDelta)} - _EXECUTABLE_DELTA_FIELDS)
-)
-
-
-class BranchParentSkillModeConflict(BranchDeltaNotSupported):
-    """A ``skill_mode`` delta forked from a parent carrying a baked-in pack.
-
-    The precondition the other skill gates cannot see: the *parent's own*
-    skill mode. A ``with-skill`` parent's ``setup()`` injects the pack into
-    the Dockerfile it builds, so the pack is in the image — and the
-    ``env-ready`` snapshot is a commit of that image. A ``no-skill`` child
-    restoring it finds ``/skills`` already there, ``deploy_skills`` correctly
-    deploys nothing on top, and the arm is reported as ``no-skill`` while the
-    agent runs *with* the parent's skills. Nothing downstream can detect that:
-    the child's config, provenance and ablation row all say ``no-skill``.
-
-    So the delta is refused unless the snapshot's world provably carries no
-    pack — i.e. the parent's own recorded skill mode is ``no-skill``, the one
-    mode whose skill policy resolves no host directory and therefore never
-    reaches ``inject_skills_into_dockerfile``. Each arm then deploys its own
-    skills (or none) at install time, over a pack-free image.
-    """
-
-
-class BranchChildExecutionNotSupported(NotImplementedError):
-    """The engine cannot run this fork's children safely (fail closed).
-
-    Raised when the *boundary*, not the delta, makes execution unsound: an
-    ``env-ready`` fork whose snapshot omits the container layer. Every child of
-    that stage re-runs ``install_agent()`` for itself (see
-    :mod:`benchflow.branch_skill`), and without the container layer the restore
-    cannot undo the previous child's installation — the second child would
-    install on top of the first and report the result as a controlled
-    comparison.
-    """
-
-
-# The per-child runner: given the child's branch node, run its continuation and
-# return the scalar return. No ``int`` index — a caller that needs per-child
-# prompts binds them into a closure (see ``run_child`` in :func:`branch`).
-ChildRunner = Callable[[RolloutNode], Awaitable[float]]
-
-
-#: The result-bearing Rollout attributes a child mutates that are not part of
-#: the linear execution state above. Derived from an audit of everything
-#: :meth:`Rollout._build_result` reads, against what a child's
-#: ``connect() -> execute() -> verify() -> disconnect()`` writes:
-#:
-#: * ``_timing`` — ``execute()`` *accumulates* ``agent_execution`` and
-#:   ``verify()`` writes into the same dict, so without this the parent's
-#:   ``timing.json`` reported the parent's time plus every arm's (observed
-#:   live by @Galius5136 on a two-arm ``pre-verify`` ablation).
-#: * ``_verifier_error`` — ``verify()`` assigns it unconditionally, so a
-#:   clean parent inherited the last child's verifier failure, and a failed
-#:   parent had its own diagnostic erased by a child that scored cleanly.
-#: * ``_diagnostics`` — ``execute()`` records prompt-timeout diagnostics and
-#:   ``verify()`` records verifier-timeout ones; a child's timeout would
-#:   surface in the parent's ``result.json`` as the parent's own.
-#: * ``_native_usage_metrics`` / ``_native_usage_checkpoint`` — ``execute()``
-#:   accumulates native ACP token usage into the first and re-bases the
-#:   second; ``cleanup()``'s ``_finalize_usage_metrics`` then promotes the
-#:   first into ``_usage_metrics``, i.e. the children's tokens are billed to
-#:   the parent. The same accumulation bug as ``_timing``, one field further
-#:   from the result.
-#:
-#: The remaining four are result-bearing and reachable from a caller-supplied
-#: ``run_child`` (which may drive any phase it likes), so they are scoped for
-#: the same reason even though the engine's own runners do not write them:
-#: ``_error``, ``_export_error``, ``_evolved_skills``, ``_usage_metrics``.
-#:
-#: Audited and deliberately *not* scoped: ``_rollout_dir`` / ``_rollout_name``
-#: / ``_resolved_prompts`` / ``_task_skill_policy`` (written by ``setup()``,
-#: which no child re-runs on the shared instance), ``_agent_name`` (written by
-#: ``connect()``, but from the parent's own config — a child cannot change what
-#: it resolves to), ``_started_at`` (the parent's start is the parent's start,
-#: and rolling it back per child would be wrong anyway), ``_terminal_timeout``
-#: (only ``_record_agent_timeout`` writes it, which lives in ``run()``), and
-#: the ``_provider_*_cached`` trio (written by ``cleanup()`` only). A
-#: fresh-rollout child touches none of these: it drives a Rollout of its own.
-_RESULT_STATE_FIELDS: tuple[str, ...] = (
-    "_timing",
-    "_verifier_error",
-    "_diagnostics",
-    "_native_usage_metrics",
-    "_native_usage_checkpoint",
-    "_error",
-    "_export_error",
-    "_evolved_skills",
-    "_usage_metrics",
-)
-
-
-@dataclass
-class _LinearState:
-    """A scoped snapshot of a Rollout's linear (non-tree) execution state.
-
-    Captured before a branch child runs and restored after — this is what
-    makes a branch child an *isolated sub-rollout* rather than a re-entrant
-    mutation of the shared Rollout instance. The stage registry (RFC §3.2) is
-    scoped the same way: a child that runs through its own ``pre-verify``
-    boundary records that snapshot on its own node, and it must not become the
-    *parent's* pre-verify — a later branch there would fork the child's world.
-
-    Two kinds of state, scoped for two different reasons. The named fields are
-    the *execution* state — where the rollout is and what it has done. The
-    ``result_state`` bag is the **result-bearing** state
-    (:data:`_RESULT_STATE_FIELDS`): fields no child needs but every in-place
-    child writes, which end up in the parent's own ``result.json`` /
-    ``timing.json`` if they are left where the last child put them. An
-    in-place child continues the *shared* Rollout, so "the parent's linear
-    state is exactly what it was before" has to cover what the parent
-    *reports*, not only where its cursor sits.
-
-    Everything mutable in that bag is deep-copied in both directions.
-    ``restore_onto`` runs once per child — not only at the end — so handing a
-    child the captured dict itself would let child *k* mutate the snapshot
-    child *k+1* and the parent are both restored from.
-    """
-
-    cursor: RolloutNode
-    trajectory: list[dict]
-    n_tool_calls: int
-    phase: str
-    rewards: dict | None
-    trajectory_source: TrajectorySource | None
-    partial_trajectory: bool
-    session_tool_count: int
-    session_traj_count: int
-    executed_prompts: list[str]
-    stage_snapshots: dict[str, StageSnapshot]
-    stage_nodes: dict[str, RolloutNode]
-    result_state: dict[str, Any]
-
-    @classmethod
-    def capture(cls, rollout: Rollout) -> _LinearState:
-        """Snapshot ``rollout``'s linear state — a shallow copy of the trajectory."""
-        return cls(
-            cursor=rollout._cursor,
-            trajectory=list(rollout._trajectory),
-            n_tool_calls=rollout._n_tool_calls,
-            phase=rollout._phase,
-            rewards=copy.deepcopy(rollout._rewards),
-            trajectory_source=rollout._trajectory_source,
-            partial_trajectory=rollout._partial_trajectory,
-            session_tool_count=getattr(rollout, "_session_tool_count", 0),
-            session_traj_count=getattr(rollout, "_session_traj_count", 0),
-            executed_prompts=list(rollout._executed_prompts),
-            stage_snapshots=dict(getattr(rollout, "_stage_snapshots", {})),
-            stage_nodes=dict(getattr(rollout, "_stage_nodes", {})),
-            # hasattr, not a default: tests build a Rollout through
-            # ``__new__()`` (the established pattern in rollout.py), and
-            # inventing an attribute the instance never had would be a
-            # different kind of mutation. What is absent stays absent.
-            result_state={
-                name: copy.deepcopy(getattr(rollout, name))
-                for name in _RESULT_STATE_FIELDS
-                if hasattr(rollout, name)
-            },
-        )
-
-    def restore_onto(self, rollout: Rollout) -> None:
-        """Write this snapshot back onto ``rollout`` — undoing a child's mutations."""
-        rollout._cursor = self.cursor
-        rollout._trajectory = list(self.trajectory)
-        rollout._n_tool_calls = self.n_tool_calls
-        rollout._phase = self.phase
-        rollout._rewards = copy.deepcopy(self.rewards)
-        rollout._trajectory_source = self.trajectory_source
-        rollout._partial_trajectory = self.partial_trajectory
-        rollout._session_tool_count = self.session_tool_count
-        rollout._session_traj_count = self.session_traj_count
-        rollout._executed_prompts = list(self.executed_prompts)
-        rollout._stage_snapshots = dict(self.stage_snapshots)
-        rollout._stage_nodes = dict(self.stage_nodes)
-        for name, value in self.result_state.items():
-            setattr(rollout, name, copy.deepcopy(value))
-
-
-def _resolve_layers(snapshot_layers: Iterable[str], *, subject: str) -> frozenset[str]:
-    """Validate a layer set — the one gate every snapshot path uses.
-
-    ``subject`` names the operation in the diagnostic ("branch", "stage
-    snapshot 'env-ready'", "branch_at_stage('pre-verify'): the recorded stage
-    snapshot").
-
-    Called on *requested* layers by the cursor branch and by
-    :func:`capture_stage`, and on the layers a stage branch *derived* from its
-    recorded snapshot — the empty set is the same fail-open either way, so it
-    is rejected in one place rather than trusted because it came from a
-    checkpoint that was already taken.
-    """
-    layers = frozenset(snapshot_layers)
-    unknown = layers - _SNAPSHOT_LAYERS
-    if unknown:
-        raise ValueError(
-            f"unknown snapshot_layers {sorted(unknown)!r} — allowed layers "
-            f"are {sorted(_SNAPSHOT_LAYERS)!r}"
-        )
-    if not layers:
-        raise ValueError(
-            f"{subject} needs at least one layer — a fork whose checkpoint "
-            "captures nothing has no roll-back point, so nothing is restored "
-            "between children and every child runs in the world the previous "
-            "one left"
-        )
-    return layers
-
-
-def _gate_layers(
-    rollout: Rollout,
-    layers: frozenset[str],
-    *,
-    subject: str,
-    requested: str,
-    gate_sandbox: bool = False,
-) -> tuple[Any, Any]:
-    """Fail closed when a requested layer's plane or capability is missing.
-
-    Returns the ``(environment, sandbox)`` live objects for the requested
-    layers — ``None`` for a layer that was not requested, which is exactly the
-    "layer not requested" signal :func:`checkpoint_composed` /
-    :func:`restore_composed` take. Every diagnostic names ``subject``, so a
-    stage capture says *which stage* could not be taken (RFC §3.2) instead of
-    the generic branch message.
-
-    The container layer keeps its #384 semantics: the Branch lifecycle
-    composes container ⊃ environment-state ⊃ agent-session, and restoring only
-    environment state can produce inconsistent state for runs that mutate
-    process/service state the Environment manifest does not capture — so a
-    missing ``supports_snapshot`` never silently degrades.
-    """
-    if "environment" in layers and getattr(rollout, "_environment", None) is None:
-        raise RuntimeError(
-            f"{subject} needs the Environment plane — there is no world to "
-            "snapshot. Pass RolloutConfig(environment_manifest=...)."
-        )
-    sandbox = getattr(rollout, "_env", None)
-    if (gate_sandbox or "sandbox" in layers) and not getattr(
-        sandbox, "supports_snapshot", False
-    ):
-        sandbox_name = type(sandbox).__name__ if sandbox else "<none>"
-        raise RuntimeError(
-            f"{subject} cannot run with {requested}: the active "
-            f"sandbox {sandbox_name!r} does not implement container-level "
-            "snapshot/restore. Use a provider whose Sandbox satisfies the "
-            "checkpoint contract (DockerSandbox or DaytonaSandbox in direct "
-            "mode), or drop the sandbox layer if Environment-state "
-            "checkpoint is sufficient for this run."
-        )
-    return (
-        rollout._environment if "environment" in layers else None,
-        sandbox if "sandbox" in layers else None,
-    )
-
-
-def _validate_skill_delta(
-    rollout: Rollout,
-    skill_mode: str,
-    *,
-    index: int,
-    at_stage: str | None,
-    layers: frozenset[str],
-    run_child: ChildRunner | None,
-) -> None:
-    """Gate a ``skill_mode`` delta before anything is quiesced (RFC §3.3).
-
-    Four preconditions, each a way the delta would otherwise measure nothing:
-
-    * **the stage** — skills are deployed by ``install_agent()``, so only the
-      ``env-ready`` boundary (captured before it) can vary them; a cursor
-      branch forks a world that already resolved the question.
-    * **the container layer** — skills live in the container filesystem, so an
-      environment-state-only checkpoint cannot roll the parent's pack back and
-      a ``no-skill`` child would still find it mounted.
-    * **the parent's own skill mode** — the layer above rolls the container
-      back *to the parent's env-ready image*, and a ``with-skill`` parent
-      baked the pack into that image at build time. Rolling back to it hands
-      a ``no-skill`` child the pack it is supposed to be running without, and
-      every artifact still labels the arm ``no-skill``. Only a parent whose
-      recorded mode is ``no-skill`` provably carries no pack, so anything else
-      raises :class:`BranchParentSkillModeConflict`.
-    * **the runner** — a caller-supplied ``run_child`` owns the child's
-      execution, so it, not the engine, would decide what the child installs.
-
-    The task's own skill policy is resolved here too, so ``with-skill``
-    against a task shipping no bundled pack raises the same typed error
-    ``setup()`` would raise, before the branch restored anything.
-    """
-    where = f"stage {at_stage!r}" if at_stage is not None else "the cursor"
-    if at_stage != SKILL_DELTA_STAGE:
-        raise BranchDeltaNotSupported(
-            f"deltas[{index}].skill_mode={skill_mode!r} executes only at "
-            f"the {SKILL_DELTA_STAGE!r} stage boundary, but this branch forks "
-            f"from {where}: skills are deployed by install_agent(), which has "
-            "already run by then. Capture the boundary with "
-            "RolloutConfig(snapshot_stages={'env-ready'}, "
-            "snapshot_layers={'environment', 'sandbox'}) and fork with "
-            "branch_at_stage('env-ready', ...) — the child then runs as a "
-            "fresh rollout over the restored sandbox (use_prebuilt_env)."
-        )
-    if SKILL_DELTA_LAYER not in layers:
-        raise BranchDeltaNotSupported(
-            f"deltas[{index}].skill_mode={skill_mode!r} needs the "
-            f"{SKILL_DELTA_LAYER!r} layer in the {SKILL_DELTA_STAGE!r} "
-            f"snapshot, which captured layers={sorted(layers)!r}: skills are "
-            "deployed into the container filesystem, and an "
-            "environment-state-only checkpoint cannot roll them back — the "
-            "child would re-install against the parent's skills. Re-capture "
-            "the stage with snapshot_layers={'environment', 'sandbox'} (or "
-            "{'sandbox'} alone for a stateless environment)."
-        )
-    parent_mode = rollout._config.recorded_skill_mode
-    if parent_mode != SKILL_MODE_NO_SKILL:
-        raise BranchParentSkillModeConflict(
-            f"deltas[{index}].skill_mode={skill_mode!r} cannot fork a parent "
-            f"whose own skill mode is {parent_mode!r}: only a "
-            f"{SKILL_MODE_NO_SKILL!r} parent provably bakes no skill pack "
-            f"into the image its setup() builds, and the "
-            f"{SKILL_DELTA_STAGE!r} snapshot every child restores is a commit "
-            "of that image. A 'no-skill' child of a parent that did bake one "
-            "restores the pack, deploys nothing on top of it, runs *with* the "
-            "parent's skills, and is still recorded as 'no-skill' in its "
-            "config, its provenance and its ablation row. Run the parent with "
-            f"skill_mode={SKILL_MODE_NO_SKILL!r} and let each arm's delta "
-            "deploy its own skills at install time."
-        )
-    if run_child is not None:
-        raise ValueError(
-            f"deltas[{index}].skill_mode cannot be combined with an explicit "
-            "run_child — the caller's runner owns the child's execution, so "
-            "the engine cannot re-run install_agent() under the switched "
-            "mode. Drop run_child so the fresh-rollout runner executes the "
-            "delta, or switch skills inside your own runner."
-        )
-    resolve_child_skill_policy(rollout._config, skill_mode)
-
-
-def _validate_config_delta(
-    overlay: dict[str, Any],
-    *,
-    index: int,
-    at_stage: str | None,
-    run_child: ChildRunner | None,
-) -> None:
-    """Gate a ``config_override`` delta before anything is quiesced (RFC §3.3).
-
-    Three preconditions, each a way the delta would otherwise measure nothing
-    (or worse, measure the wrong thing):
-
-    * **the stage** — the C-axis overlay is deep-merged into the task's
-      resolved config by ``setup()``, which only a fresh child rollout re-runs;
-      by any later boundary the parent's config has already been consumed
-      (its timeout enforced, its prompts resolved, its sandbox built), so a
-      child there would record the override and run without it.
-    * **the allowlist** — the same fail-closed allowlist the run-level overlay
-      passes through (#790): scorer-touching sections (``verifier`` /
-      ``reward`` / ``solution`` / ``oracle``) are rejected *here*, before the
-      parent is quiesced, not at child setup after a snapshot was restored.
-    * **the runner** — a caller-supplied ``run_child`` owns the child's
-      execution, so the engine cannot run the child under the merged config.
-    """
-    where = f"stage {at_stage!r}" if at_stage is not None else "the cursor"
-    if at_stage != FRESH_CHILD_STAGE:
-        raise BranchDeltaNotSupported(
-            f"deltas[{index}].config_override executes only at the "
-            f"{FRESH_CHILD_STAGE!r} stage boundary, but this branch forks from "
-            f"{where}: the overlay is deep-merged into the task's resolved "
-            "config by the child's own setup(), which only a fresh child "
-            "rollout (use_prebuilt_env) re-runs — by any later boundary the "
-            "parent's config has already been consumed. Capture the boundary "
-            "with RolloutConfig(snapshot_stages={'env-ready'}, "
-            "snapshot_layers={'environment', 'sandbox'}) and fork with "
-            "branch_at_stage('env-ready', ...)."
-        )
-    if run_child is not None:
-        raise ValueError(
-            f"deltas[{index}].config_override cannot be combined with an "
-            "explicit run_child — the caller's runner owns the child's "
-            "execution, so the engine cannot run setup() under the merged "
-            "config. Drop run_child so the fresh-rollout runner executes the "
-            "delta, or apply the override inside your own runner."
-        )
-    try:
-        validate_overlay(overlay)
-    except ValueError as exc:
-        raise ValueError(f"deltas[{index}].config_override: {exc}") from None
-
-
-def _validate_environment_delta(
-    rollout: Rollout,
-    environment_ref: str,
-    *,
-    index: int,
-    at_stage: str | None,
-    run_child: ChildRunner | None,
-) -> None:
-    """Gate an ``environment_ref`` delta before anything is quiesced (RFC §3.3).
-
-    The stage and runner gates mirror the other fresh-child deltas: the child
-    manifest binds at *provision* time, and provisioning happens only on the
-    fresh-child path a ``env-ready`` fork runs (an in-place child at a later
-    boundary would keep the parent's provisioned services no matter what its
-    delta records). The delta's own content gates — a manifest-bound parent,
-    a resolvable ref, the same image, framework-started services on both
-    sides — live in :func:`~benchflow.branch_skill.resolve_environment_ref_delta`,
-    which the child runner re-derives its manifest through, so validation and
-    execution cannot drift.
-    """
-    where = f"stage {at_stage!r}" if at_stage is not None else "the cursor"
-    if at_stage != FRESH_CHILD_STAGE:
-        raise BranchDeltaNotSupported(
-            f"deltas[{index}].environment_ref={environment_ref!r} executes "
-            f"only at the {FRESH_CHILD_STAGE!r} stage boundary, but this "
-            f"branch forks from {where}: the manifest's environment plane is "
-            "provisioned over the restored sandbox by the fresh child rollout "
-            "(use_prebuilt_env), and at any later boundary the parent's "
-            "provisioned services survive the fork — the delta would be "
-            "recorded but not enforced. Capture the boundary with "
-            "RolloutConfig(snapshot_stages={'env-ready'}, "
-            "snapshot_layers={'environment', 'sandbox'}) and fork with "
-            "branch_at_stage('env-ready', ...)."
-        )
-    if run_child is not None:
-        raise ValueError(
-            f"deltas[{index}].environment_ref cannot be combined with an "
-            "explicit run_child — the caller's runner owns the child's "
-            "execution, so the engine cannot provision the child manifest's "
-            "services. Drop run_child so the fresh-rollout runner executes "
-            "the delta, or provision inside your own runner."
-        )
-    resolve_environment_ref_delta(
-        rollout._config.environment_manifest,
-        environment_ref,
-        subject=f"deltas[{index}].environment_ref",
-    )
-
-
-def _runs_fresh_children(at_stage: str | None, run_child: ChildRunner | None) -> bool:
-    """Whether the engine will run this fork's children as fresh rollouts.
-
-    True for every engine-run child of ``env-ready``, whatever its delta. That
-    boundary precedes ``install_agent()``, so the restored world has no agent
-    binary, no sandbox user, no seeded verifier workspace, no path lockdown and
-    no skill pack: an in-place child there connects to something the restore
-    deleted, or — when the agent happens to survive in the base image — scores
-    a world missing everything installation deploys and reports it as an
-    ordinary child under one recorded delta. A caller-supplied ``run_child``
-    owns its own execution and is left alone.
-    """
-    return run_child is None and at_stage == FRESH_CHILD_STAGE
-
-
-def _validate_fresh_children(layers: frozenset[str], *, at_stage: str) -> None:
-    """Gate a fresh-child fork before anything is quiesced.
-
-    One precondition beyond the per-delta gates: the stage snapshot must carry
-    the container layer. Every child re-runs ``install_agent()``, which writes
-    to the container filesystem, so an environment-state-only checkpoint cannot
-    put the world back between children — child *k+1* would install on top of
-    child *k*'s agent, user, lockdown and skill pack. The skill-delta gate says
-    the same thing in skill terms and fires first when a skill delta is
-    present; this covers the children that carry no skill delta at all.
-    """
-    if FRESH_CHILD_LAYER not in layers:
-        raise BranchChildExecutionNotSupported(
-            f"a branch at the {at_stage!r} stage boundary runs every child as "
-            f"a fresh rollout, which needs the {FRESH_CHILD_LAYER!r} layer in "
-            f"that snapshot; it captured layers={sorted(layers)!r}. "
-            f"{at_stage!r} precedes install_agent(), so each child installs "
-            "the agent (and its skills, user and lockdown) for itself, and an "
-            "environment-state-only checkpoint cannot roll that back between "
-            "children. Re-capture the stage with "
-            "snapshot_layers={'environment', 'sandbox'} (or {'sandbox'} alone "
-            "for a stateless environment), or supply your own run_child if "
-            "you are executing the children yourself."
-        )
-
-
-async def capture_stage(
-    rollout: Rollout,
-    stage: str,
-    *,
-    snapshot_layers: Iterable[str] = frozenset({"environment"}),
-) -> StageSnapshot:
-    """Take the composed stage snapshot at ``rollout``'s cursor (RFC §3.2).
-
-    The stage-boundary policy's one capture path: validate the stage name and
-    the requested layers, fail closed on a missing plane or capability (the
-    diagnostic names the stage), compose the checkpoint on the cursor node —
-    so the stage tag flows into ``tree.json`` through the node's own recorded
-    state — and register it under ``stage`` on the rollout. The registry is
-    re-serialized to ``stage_snapshots.json`` after every capture, with write
-    failures logged and isolated from the rollout.
-
-    Callers are the lifecycle's own boundaries (``start()``/``verify()``, gated
-    on ``RolloutConfig.snapshot_stages``) and ``Rollout.mark_stage`` for the
-    boundaries only the caller can see. Two stages captured without an
-    intervening Step land on the same node (``pre-verify`` and ``post-verify``
-    of a rollout that verifies once, say): the registry keeps both, and the
-    node's own tag is the most recent — a node has one checkpoint, and the
-    branch that adopts it re-tags it with the stage it branched from.
-    """
-    validate_stage(stage)
-    subject = f"stage snapshot {stage!r}"
-    layers = _resolve_layers(snapshot_layers, subject=subject)
-    snap_env, snap_sandbox = _gate_layers(
-        rollout,
-        layers,
-        subject=subject,
-        requested=f"snapshot_layers={sorted(layers)!r}",
-    )
-    try:
-        snapshot = await _checkpoint_composed(
-            rollout._cursor,
-            environment=snap_env,
-            sandbox=snap_sandbox,
-            stage=stage,
-        )
-    except Exception as exc:
-        exc.add_note(
-            f"stage snapshot {stage!r} (snapshot_layers={sorted(layers)!r}) "
-            "failed during checkpoint — the rollout fails closed; no partial "
-            "checkpoint was recorded on the node."
-        )
-        raise
-    rollout._stage_snapshots[stage] = snapshot
-    rollout._stage_nodes[stage] = rollout._cursor
-
-    run_dir = getattr(rollout, "_rollout_dir", None)
-    if run_dir is not None:
-        try:
-            write_stage_snapshots(run_dir=run_dir, snapshots=rollout._stage_snapshots)
-        except Exception:
-            logger.warning(
-                "stage snapshot artifact write under %s failed — the recorded "
-                "stage registry is unaffected",
-                run_dir,
-                exc_info=True,
-            )
-    return snapshot
-
-
-def _recorded_stage_checkpoint(
-    rollout: Rollout, stage: str, snapshot_layers: Iterable[str] | None
-) -> tuple[StageSnapshot, RolloutNode, frozenset[str]]:
-    """Resolve a recorded stage into ``(snapshot, branch node, layers)``.
-
-    A stage that was never captured fails closed with
-    :class:`~benchflow.branch_stage.BranchStageNotCaptured` naming what *was*
-    captured — degrading to a checkpoint at the cursor would fork a different
-    world state than the caller asked for and mislabel it in provenance. The
-    layers are derived from the recorded snapshot; an explicit
-    ``snapshot_layers`` that disagrees is a caller bug, not a request to
-    re-snapshot.
-
-    Derived is not the same as trusted: a snapshot carrying *neither* ref
-    derives the empty set, which the layer gates below then have nothing to
-    check — the fork would restore nothing between children and still record a
-    clean stage fork. So the derived set goes through :func:`_resolve_layers`,
-    the same non-empty gate the cursor branch runs on its requested set, and
-    it runs before the disagreement check: an empty capture is broken however
-    it is described, and ``snapshot_layers=set()`` would otherwise "agree"
-    with it.
-    """
-    validate_stage(stage)
-    registry: dict[str, StageSnapshot] = getattr(rollout, "_stage_snapshots", {})
-    snapshot = registry.get(stage)
-    if snapshot is None:
-        raise BranchStageNotCaptured(
-            f"no snapshot recorded at stage {stage!r} — this rollout captured "
-            f"{captured_stages(registry)!r}. Request the stage up front with "
-            "RolloutConfig(snapshot_stages={...}), or record it at the cut "
-            "point with Rollout.mark_stage()."
-        )
-    layers = _resolve_layers(
-        (
-            layer
-            for layer, ref in (
-                ("environment", snapshot.environment_ref),
-                ("sandbox", snapshot.sandbox_ref),
-            )
-            if ref is not None
-        ),
-        subject=f"branch_at_stage({stage!r}): the recorded stage snapshot",
-    )
-    if snapshot_layers is not None and frozenset(snapshot_layers) != layers:
-        raise ValueError(
-            f"branch_at_stage({stage!r}, snapshot_layers="
-            f"{sorted(frozenset(snapshot_layers))!r}) disagrees with the "
-            f"layers stage {stage!r} actually captured ({sorted(layers)!r}) — "
-            "a stage branch restores exactly what the stage snapshotted; "
-            "re-run with the layers the capture used, or capture the stage "
-            "with the layers you want."
-        )
-    node = getattr(rollout, "_stage_nodes", {}).get(stage, rollout._cursor)
-    return snapshot, node, layers
+#: Back-compat alias — tests and callers imported the scoped-state class under
+#: its original private name.
+_LinearState = LinearState
 
 
 async def branch(
@@ -767,28 +160,9 @@ async def branch(
 
     ``deltas`` records the exactly-one-controlled-change each child runs
     under (RFC §3.3) — one :class:`BranchDelta` (or ``None`` = zero delta)
-    per child, validated before anything runs. Three fields execute.
-    ``injected_prompt`` becomes the child's continuation prompt, delivered as
-    the child's user-visible first message and recorded in provenance as a
-    content hash. ``skill_mode`` re-runs ``install_agent()`` under the switched
-    mode — the with-skill / no-skill ablation; it requires
-    ``at_stage="env-ready"`` with the container layer in that snapshot **and a
-    parent whose own recorded skill mode is ``no-skill``** (a with-skill
-    parent baked its pack into the image the snapshot commits, so a no-skill
-    child would restore it and be mislabeled), and fails closed otherwise.
-    ``config_override`` runs the child as a fresh rollout whose config is the
-    parent's with the allowlisted C-axis overlay deep-merged on top (#790
-    machinery — same allowlist, same content addressing); it executes only
-    from ``at_stage="env-ready"`` (the child's own ``setup()`` is what applies
-    the overlay) and a non-allowlisted key fails closed before any child runs.
-    ``environment_ref`` swaps the child onto a different environment-registry
-    manifest — the *service-level* slice only (same image, different
-    framework-started services: the ``env0@prod`` vs ``env0@outage`` pattern);
-    the fresh child provisions the child manifest's services over the restored
-    container, and a manifest that changes the image or an entrypoint-owned
-    lifecycle fails closed
-    (:class:`~benchflow.branch_skill.BranchEnvironmentImageConflict` /
-    :class:`BranchDeltaNotSupported`) before any child runs.
+    per child, validated before anything runs (see
+    :func:`~benchflow.branch_policy.validate_deltas` for the per-field gates
+    and :mod:`benchflow.branch_children` for how each executable field runs).
 
     **How a child is executed depends on the boundary, not on the delta.**
     ``at_stage="env-ready"`` precedes ``install_agent()``, so every child the
@@ -799,8 +173,9 @@ async def branch(
     deleted, or score a world missing everything installation deploys and
     report it as an ordinary single-delta child; that fork therefore requires
     the container layer and raises
-    :class:`BranchChildExecutionNotSupported` without it. At every other
-    boundary the agent is already installed and children run in place.
+    :class:`~benchflow.branch_policy.BranchChildExecutionNotSupported` without
+    it. At every other boundary the agent is already installed and children
+    run in place.
 
     After this returns, ``rollout``'s linear state is exactly what it was
     before — the tree gained ``n`` children at the cursor, nothing else
@@ -826,81 +201,27 @@ async def branch(
     """
     stage_snapshot: StageSnapshot | None = None
     if at_stage is not None:
-        stage_snapshot, stage_node, layers = _recorded_stage_checkpoint(
+        stage_snapshot, stage_node, layers = recorded_stage_checkpoint(
             rollout, at_stage, snapshot_layers
         )
         subject = f"branch_at_stage({at_stage!r})"
     else:
         subject = "branch"
-        layers = _resolve_layers(
+        layers = resolve_layers(
             frozenset({"environment"}) if snapshot_layers is None else snapshot_layers,
             subject=subject,
         )
     if n < 2:
         raise ValueError(f"a branch forks into >= 2 children, got n={n}")
 
-    # Validate the whole delta vector before anything runs — a bad or
-    # not-yet-executable delta fails closed with nothing quiesced,
-    # checkpointed, or run.
-    if deltas is not None:
-        if len(deltas) != n:
-            raise ValueError(
-                f"deltas must carry exactly one entry per child: got "
-                f"{len(deltas)} deltas for n={n}"
-            )
-        for index, delta in enumerate(deltas):
-            if delta is None:
-                continue
-            for field_name in _UNSUPPORTED_DELTA_FIELDS:
-                if getattr(delta, field_name) is not None:
-                    raise BranchDeltaNotSupported(
-                        f"deltas[{index}].{field_name} is set, but the branch "
-                        "engine executes injected_prompt and skill_mode only "
-                        f"— {field_name!r} runs in the rollout-branching RFC "
-                        "follow-on (child-as-fresh-rollout via "
-                        "use_prebuilt_env). The field is already recorded in "
-                        "the schema and provenance; drop it to branch today."
-                    )
-            if delta.config_override is not None:
-                _validate_config_delta(
-                    delta.config_override,
-                    index=index,
-                    at_stage=at_stage,
-                    run_child=run_child,
-                )
-            if delta.environment_ref is not None:
-                _validate_environment_delta(
-                    rollout,
-                    delta.environment_ref,
-                    index=index,
-                    at_stage=at_stage,
-                    run_child=run_child,
-                )
-            skill_mode = delta.skill_mode
-            if skill_mode is not None:
-                _validate_skill_delta(
-                    rollout,
-                    skill_mode,
-                    index=index,
-                    at_stage=at_stage,
-                    layers=layers,
-                    run_child=run_child,
-                )
-            if delta.injected_prompt is not None and run_child is not None:
-                raise ValueError(
-                    f"deltas[{index}].injected_prompt cannot be combined "
-                    "with an explicit run_child — the caller's runner owns "
-                    "the child's prompts. Bind the prompt into the "
-                    "run_child closure, or drop run_child so the default "
-                    "runner delivers it."
-                )
-
-    # The boundary's own gate: at ``env-ready`` every engine-run child is a
-    # fresh rollout (whatever its delta, including no delta), so the container
-    # layer is required even for a fork carrying no deltas at all.
-    fresh_children = _runs_fresh_children(at_stage, run_child)
+    # Admit the request — the whole delta vector and the boundary's own gate
+    # fail closed with nothing quiesced, checkpointed, or run.
+    validate_deltas(
+        rollout, deltas, n=n, at_stage=at_stage, layers=layers, run_child=run_child
+    )
+    fresh_children = runs_fresh_children(at_stage, run_child)
     if fresh_children:
-        _validate_fresh_children(layers, at_stage=str(at_stage))
+        validate_fresh_children(layers, at_stage=str(at_stage))
 
     # Fail closed when a requested layer's plane or capability is missing
     # (#384). ``require_sandbox_snapshot`` keeps its original check-only
@@ -908,7 +229,7 @@ async def branch(
     # way and then actually composes the sandbox into the checkpoint. A stage
     # branch gates too — the planes must still be live to restore what the
     # stage captured.
-    snap_env, snap_sandbox = _gate_layers(
+    snap_env, snap_sandbox = gate_layers(
         rollout,
         layers,
         subject=subject,
@@ -920,41 +241,27 @@ async def branch(
         gate_sandbox=require_sandbox_snapshot,
     )
     composed = stage_snapshot is not None or layers != frozenset({"environment"})
-
     parent = stage_node if stage_snapshot is not None else rollout._cursor
     runner = run_child if run_child is not None else make_default_runner(rollout)
     run_dir = getattr(rollout, "_rollout_dir", None)
 
     # quiesce — pause the agent before snapshotting so the checkpoint is
-    # consistent (the Branch lifecycle quiesces first).
+    # consistent (the Branch lifecycle quiesces first); then checkpoint — the
+    # roll-back point (a stage branch adopts the snapshot the boundary took).
     await rollout.disconnect()
-
-    # checkpoint — snapshot the requested layers at the parent; the roll-back
-    # point. The default environment-only request keeps the legacy path (and
-    # the legacy bare-StateSnapshot shape on node.state["snapshot"]) intact.
-    # A stage branch adopts the snapshot the stage boundary already took —
-    # re-snapshotting there would capture a world that has since moved on.
-    if stage_snapshot is not None:
-        _adopt_checkpoint(parent, stage_snapshot)
-    else:
-        try:
-            if composed:
-                await _checkpoint_composed(
-                    parent, environment=snap_env, sandbox=snap_sandbox
-                )
-            else:
-                await _checkpoint_branch(parent, rollout._environment)
-        except Exception as exc:
-            exc.add_note(
-                f"branch(snapshot_layers={sorted(layers)!r}) failed during "
-                "checkpoint — the branch fails closed; no partial checkpoint "
-                "was recorded on the node."
-            )
-            raise
+    await checkpoint_parent(
+        rollout,
+        parent,
+        stage_snapshot=stage_snapshot,
+        composed=composed,
+        snap_env=snap_env,
+        snap_sandbox=snap_sandbox,
+        layers=layers,
+    )
 
     # The parent's linear state, captured once. Each child runs against a fresh
     # restore of this; the parent is restored to it at the end.
-    saved = _LinearState.capture(rollout)
+    saved = LinearState.capture(rollout)
 
     # The parent's *on-disk* state needs the same treatment, for the same
     # reason. Children share the parent's sandbox, and the sandbox bind-mounts
@@ -973,29 +280,32 @@ async def branch(
         if run_dir is not None
         else None
     )
-
-    children: list[RolloutNode] = []
+    transaction = BranchTransaction(
+        rollout=rollout,
+        n=n,
+        deltas=deltas,
+        parent=parent,
+        saved=saved,
+        runner=runner,
+        run_child=run_child,
+        composed=composed,
+        snap_env=snap_env,
+        snap_sandbox=snap_sandbox,
+        fresh_children=fresh_children,
+        run_dir=run_dir,
+        holder=holder,
+        # Resolved through *this* module's namespace at call time — the patch
+        # seam tests/test_branch_child_result.py pins (a fake fresh-child
+        # runner, a recording result writer).
+        fresh_runner_factory=make_fresh_child_runner,
+        write_child_result=write_in_place_child_result,
+    )
     try:
         try:
-            await _run_children(
-                rollout,
-                n=n,
-                deltas=deltas,
-                parent=parent,
-                children=children,
-                saved=saved,
-                runner=runner,
-                run_child=run_child,
-                composed=composed,
-                snap_env=snap_env,
-                snap_sandbox=snap_sandbox,
-                fresh_children=fresh_children,
-                run_dir=run_dir,
-                holder=holder,
-            )
+            await transaction.run_children()
         except BaseException:
             # The exception on its way out — a child's crash, or a custody
-            # failure _run_children itself surfaced — is the caller's
+            # failure the transaction itself surfaced — is the caller's
             # diagnosis. release() must not replace it: it records and logs a
             # failure of its own (preserving the hold directory) instead of
             # raising over the real one.
@@ -1036,27 +346,38 @@ async def branch(
                 "child failed — the parent's own result may carry the child's",
                 exc_info=True,
             )
-        _write_lineage(rollout, run_dir=run_dir, parent=parent, children=children)
+        _write_lineage(
+            rollout, run_dir=run_dir, parent=parent, children=transaction.children
+        )
         raise
 
     # restore the parent's linear state — the tree grew, nothing else moved.
     saved.restore_onto(rollout)
+    value = _aggregate_fork(parent, transaction.children)
+    rollout._phase = "branched"
+    _write_lineage(
+        rollout, run_dir=run_dir, parent=parent, children=transaction.children
+    )
+    return value
 
-    # aggregate — per-child return -> V(parent), over *this* fork's children:
-    # a stage branch point can already carry the linear continuation, whose
-    # node has no reward to average. An unscored child makes V *undefined*:
-    # averaging it in as a zero would report a number computed from a score
-    # that was never observed, so the value is left unrecorded and returned
-    # as None.
-    #
-    # "Unrecorded" has to mean *removed*, not merely "not written": a node can
-    # be branched more than once, and skipping the write would leave the
-    # previous fork's V sitting on the node — ``branch()`` returning None while
-    # ``tree.json`` publishes a number for a fork that has none.
+
+def _aggregate_fork(
+    parent: RolloutNode, children: Sequence[RolloutNode]
+) -> float | None:
+    """Per-child return -> V(parent), over *this* fork's children only.
+
+    A stage branch point can already carry the linear continuation, whose node
+    has no reward to average. An unscored child makes V *undefined*: averaging
+    it in as a zero would report a number computed from a score that was never
+    observed, so the value is left unrecorded and returned as ``None``.
+
+    "Unrecorded" has to mean *removed*, not merely "not written": a node can
+    be branched more than once, and skipping the write would leave the
+    previous fork's V sitting on the node — ``branch()`` returning None while
+    ``tree.json`` publishes a number for a fork that has none.
+    """
     unscored_children = [child for child in children if UNSCORED_KEY in child.state]
-    value: float | None
     if unscored_children:
-        value = None
         parent.state.pop("value", None)
         logger.error(
             "branch at %s left %d of %d children unscored — V(parent) is "
@@ -1065,12 +386,9 @@ async def branch(
             len(unscored_children),
             len(children),
         )
-    else:
-        value = _aggregate_branch(parent, over=children)
-        parent.state["value"] = value
-    rollout._phase = "branched"
-
-    _write_lineage(rollout, run_dir=run_dir, parent=parent, children=children)
+        return None
+    value = _aggregate_branch(parent, over=children)
+    parent.state["value"] = value
     return value
 
 
@@ -1096,6 +414,11 @@ def _write_lineage(
     per-child ``provenance.json`` / ``reward.json`` for the children that were
     attached, plus the parent's own stage registry (a child that ran through a
     stage boundary of its own already overwrote the file with its version).
+
+    Kept in this module (not :mod:`benchflow.branch_report`) so the writers
+    resolve through *this* namespace — ``tests/test_rollout_branch.py``
+    patches ``benchflow.rollout_branch.write_branch_artifacts`` to prove the
+    isolation.
     """
     if run_dir is None:
         return
@@ -1116,205 +439,3 @@ def _write_lineage(
             run_dir,
             exc_info=True,
         )
-
-
-async def _run_children(
-    rollout: Rollout,
-    *,
-    n: int,
-    deltas: Sequence[BranchDelta | None] | None,
-    parent: RolloutNode,
-    children: list[RolloutNode],
-    saved: _LinearState,
-    runner: ChildRunner,
-    run_child: ChildRunner | None,
-    composed: bool,
-    snap_env: Any,
-    snap_sandbox: Any,
-    fresh_children: bool,
-    run_dir: Any,
-    holder: MountedArtifacts | None,
-) -> None:
-    """Run the fork's ``n`` children in order, appending them to ``children``.
-
-    Split out of :func:`branch` only so the mounted-artifact custody around the
-    loop reads as one bracket instead of burying the aggregation step inside a
-    ``try``. Every child is appended to ``children`` as soon as it is attached,
-    so a caller that catches a propagating child failure still sees the nodes
-    that were forked — which is how ``bench eval ablate`` reports the arm that
-    raised and the arms that never ran.
-
-    An **in-place** child of a run-dir-bearing rollout additionally leaves its
-    own ``result.json`` set (:mod:`benchflow.branch_result`): its
-    result-bearing state is scoped to zero before it runs — so what the fields
-    hold afterwards is the child's own, never the parent's showing through —
-    and the result is synthesized from that state after the child completes,
-    *before* the next restore discards it. A fresh-rollout child writes its own
-    result already and is left alone; a child that raised ends the fork and is
-    evidenced by its ``mounted/`` archive instead.
-    """
-    in_place_results = not fresh_children and run_dir is not None
-    for index in range(n):
-        # Attach a *pending* branch-child node — its real continuation Step is
-        # filled in place by the child's first execute(), so the child's work
-        # lands on the child node, not a descendant placeholder.
-        delta = deltas[index] if deltas is not None else None
-        child = rollout._tree.attach(parent)
-        # The child's delta provenance is recorded on the node itself at fork
-        # time (``None`` = the zero delta), so lineage serialization reads it
-        # from the node — never by positional alignment — and a second
-        # branch() at the same parent can never misattribute deltas.
-        child.state["delta"] = (
-            delta if delta is not None else BranchDelta()
-        ).provenance_dict()
-        # How the delta is executed, recorded on the node for the same reason:
-        # an env-ready child re-runs installation as its own Rollout, and the
-        # artifacts must say so rather than leaving a reader to infer it from
-        # the delta — which they could not, since a zero-delta child of that
-        # boundary is a fresh rollout too.
-        if fresh_children:
-            child.state["delta_execution"] = EXECUTION_FRESH_ROLLOUT
-        children.append(child)
-
-        # restore the checkpointed layers (sandbox first, then env — the
-        # reverse of checkpoint order), reset the parent's linear state, and
-        # point the cursor at the pending child for the sub-rollout.
-        if composed:
-            await _restore_composed(parent, environment=snap_env, sandbox=snap_sandbox)
-        else:
-            await _restore_branch(parent, rollout._environment)
-        saved.restore_onto(rollout)
-        rollout._cursor = child
-        started_wall = datetime.now()
-        if in_place_results:
-            # The restore just put the *parent's* result-bearing values back;
-            # zero them so the child's result below reports only what the
-            # child itself produced. The next restore brings the parent's back.
-            scope_child_result_state(rollout)
-
-        # Pick the per-child runner (every case validated above to not combine
-        # with an explicit run_child). A child forked from ``env-ready`` runs
-        # as a fresh Rollout over the just-restored sandbox — that boundary
-        # precedes install_agent(), so there is no agent in the restored world
-        # to connect to and no installed skills to keep; the child re-runs
-        # installation for itself, under the delta's skill_mode when it has
-        # one and under the parent's own recorded mode when it does not.
-        # Everywhere else the agent is already installed and the child runs in
-        # place, with an injected-prompt delta binding the child's continuation
-        # prompt into a per-child default runner — the formalized version of
-        # the caller's per-child-prompt closure.
-        child_runner = runner
-        if fresh_children:
-            child_runner = make_fresh_child_runner(
-                rollout,
-                delta=delta,
-                parent=parent,
-                branch_stage=FRESH_CHILD_STAGE,
-                run_dir=run_dir,
-            )
-        elif (
-            run_child is None
-            and delta is not None
-            and delta.injected_prompt is not None
-        ):
-            child_runner = make_default_runner(rollout, prompts=[delta.injected_prompt])
-
-        started = time.monotonic()
-        unscored: str | None = None
-        try:
-            ret = await child_runner(child)
-        except UnscoredChildError as exc:
-            # The child ran but was never scored. Record *why*, leave the
-            # node's reward unset, and keep going: the next child restores the
-            # checkpoint for itself, so one lost score does not cost the fork.
-            unscored = exc.reason
-        finally:
-            # Recorded in a finally so a child that raised still reports what
-            # it cost before it did — and, for the same reason, a child that
-            # raised still hands off whatever it wrote to the shared mounts
-            # (a crashed child's verifier output is often the only evidence of
-            # why it crashed). Handing off also empties the mounts, so the
-            # next child cannot inherit this one's files. hand_off never
-            # raises here — a raise inside this finally would replace the
-            # child's own exception — it records the failure on the holder,
-            # and raise_pending() below surfaces it once the child's outcome
-            # is safely on the node.
-            child.state[CHILD_WALL_CLOCK_KEY] = time.monotonic() - started
-            if holder is not None and run_dir is not None:
-                holder.hand_off(child_mount_dir(run_dir, parent.id, child.id))
-        if unscored is not None:
-            child.state[UNSCORED_KEY] = unscored
-            logger.error("branch child %s is unscored: %s", child.id, unscored)
-        else:
-            child.state["reward"] = float(ret)
-        if in_place_results:
-            # The instance still carries the child's own state (the restore
-            # happens on the next iteration / at the end of the fork), so this
-            # is the one window where a first-class result can be built for an
-            # in-place child. Best-effort by contract: never raises.
-            write_in_place_child_result(
-                rollout,
-                parent=parent,
-                child=child,
-                run_dir=run_dir,
-                started_at=started_wall,
-                base_trajectory_len=len(saved.trajectory),
-                base_prompt_count=len(saved.executed_prompts),
-                base_tool_calls=saved.n_tool_calls,
-            )
-        if holder is not None:
-            # Custody fails closed: a hand-off failure means the mounts may
-            # still carry this child's files, which the next child would both
-            # inherit and destroy. The child's own outcome (reward / unscored
-            # reason / result.json) is recorded above, so raising here loses
-            # no observation — the fork stops before evidence can cross arms.
-            holder.raise_pending()
-
-
-def make_default_runner(
-    rollout: Rollout, *, prompts: list[str] | None = None
-) -> ChildRunner:
-    """Build the default per-child runner bound to ``rollout``.
-
-    The default runner re-runs the child from the parent's env checkpoint with
-    a *fresh agent session* — agent-session snapshot is the unsolved hard part
-    (``docs/architecture.md``, "The hard part"), so the agent restarts per
-    child. Each child connects a fresh agent and disconnects it at the end, so
-    no two children's agents overlap (the next child connects only after the
-    previous one disconnected). ``verify()`` returning ``None``, an empty dict,
-    or a dict with a ``None`` reward raises
-    :class:`~benchflow.branch.UnscoredChildError`: the child ran but was never
-    scored, and a fabricated ``0.0`` would read as a real failing score.
-
-    ``prompts`` — the child's continuation prompts; ``None`` keeps the
-    rollout's resolved prompts. An ``injected_prompt`` delta (RFC §3.3) binds
-    here, so the injection is the child's user-visible first message — never
-    silently merged into other prompt content (#908).
-    """
-
-    async def _runner(child: RolloutNode) -> float:
-        await rollout.connect()
-        # Fill the pending branch-child node in place — the continuation Step
-        # lands on `child` itself, no content-free placeholder.
-        await rollout.execute(prompts, node=child)
-        rewards = await rollout.verify()
-        await rollout.disconnect()
-        if not rewards or rewards.get("reward") is None:
-            raise UnscoredChildError(
-                f"branch child {child.id} produced no verifier reward "
-                f"(verify() returned {rewards!r})" + _verifier_error_suffix(rollout)
-            )
-        return float(rewards["reward"])
-
-    return _runner
-
-
-def _verifier_error_suffix(rollout: Rollout) -> str:
-    """`` — <verifier error>`` when the rollout recorded one, else ``''``.
-
-    The verifier's own diagnostic is the difference between "the agent scored
-    nothing" and "the score never reached the host", so it is carried into the
-    unscored reason verbatim rather than left in the log.
-    """
-    error = getattr(rollout, "_verifier_error", None)
-    return f" — {error}" if error else ""
